Move ExtCns and related definitions to Verifier.SAW.Name.
